backend/requirements.txt
torchaudio==2.4.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Pillow==9.3.0
PyMySQL==1.0.2
Authlib==1.1.0
reportlab==3.6.12
Werkzeug==2.2.2
Pillow
requests
transformers
mutagen
matplotlib
librosa
git+https://github.com/openai/whisper.git
git+https://github.com/juanmc2005/diart.git
mysql-connector-python
python-dotenv
groq
langchain_groq
langchain